requirements.txt
fastapi
uvicorn[standard]
requests
streamlit
python-dotenv
google-generativeai
